add: 글 삭제시 확인

+++ src/page/particularContent/PostContent/index.js
@@ -2,27 +2,32 @@ import * as S from './style.ts'
 import titleImg from "../../../assets/titleImg1.svg";
 import Rectangle from "../../../components/Button/Rectangle";
 import CommentList from "./CommentList";
-import {useEffect, useState} from "react";
 import {useNavigate, useParams} from "react-router-dom";
 import InputText from "../../../components/InputText";
+import {useRecoilValue} from "recoil";
+import {authAtom} from "../../../recoil/authAtom";
 
 function PostContent({data}) {
     const navigate = useNavigate();
     const {id} = useParams();
-    const userId = localStorage.getItem("accessToken");
-    const [isWriter, setIsWriter] = useState(false);
+    const getAuth = useRecoilValue(authAtom);
+    const userId = getAuth.username||null;
+    const isManager = localStorage.getItem("isManager");
     const delPost = async () => {
-        try{
-            const response = await fetch(`/community/doc/${id}`, {
-                method:'DELETE',
-            })
-            const data = await response.json();
-            if(data.status===200){
-                console.log("글 삭제 성공");
+        if(window.confirm("정말 삭제하시겠습니까?")){
+            try{
+                const response = await fetch(`/community/doc/${id}`, {
+                    method:'DELETE',
+                })
+                const data = await response.json();
+                if(data.status===200){
+                    console.log("글 삭제 성공");
+                    navigate("/");
+                }
+            }catch (error){
+                console.log("error on : ",error);
+                navigate("/error");
             }
-        }catch (error){
-            console.log("error on : ",error);
-            navigate("/error");
         }
     }
 
@@ -31,7 +36,7 @@ function PostContent({data}) {
             <S.Header>
                 <S.HeaderHead>
                     <S.category>{data["category"]}</S.category>
-                    <S.ManagePost><S.ManageBtn onClick={delPost}>제거</S.ManageBtn> - <S.ManageBtn>수정</S.ManageBtn></S.ManagePost>
+                    {userId===id? <S.ManagePost><S.ManageBtn onClick={delPost}>제거</S.ManageBtn> - <S.ManageBtn onClick={navigate('/')}>수정</S.ManageBtn></S.ManagePost>:null}
                 </S.HeaderHead>
                 <S.titleWrap><S.titleImg src={titleImg}/><S.title>{data["title"]}</S.title></S.titleWrap>
                 <S.postInfo>{data["createdAt"]} - {data["userId"]}</S.postInfo>
